HW2 -> controller, service, repository layers implemented;

+++ src/main/java/ru/javawebinar/topjava/repository/mock/InMemoryMealRepositoryImpl.java
@@ -1,46 +1,93 @@
 package ru.javawebinar.topjava.repository.mock;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.stereotype.Repository;
 import ru.javawebinar.topjava.model.Meal;
 import ru.javawebinar.topjava.repository.MealRepository;
+import ru.javawebinar.topjava.to.MealWithExceed;
 import ru.javawebinar.topjava.util.MealsUtil;
 
-import java.util.Collection;
+import java.time.LocalTime;
+import java.util.List;
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.atomic.AtomicInteger;
+import java.util.stream.Collectors;
 
+@Repository
 public class InMemoryMealRepositoryImpl implements MealRepository {
+    private static final Logger log = LoggerFactory.getLogger(InMemoryMealRepositoryImpl.class);
     private Map<Integer, Meal> repository = new ConcurrentHashMap<>();
     private AtomicInteger counter = new AtomicInteger(0);
 
     {
-        MealsUtil.MEALS.forEach(this::save);
+        MealsUtil.MEALS.forEach(meal -> this.save(meal, 1));
     }
 
     @Override
-    public Meal save(Meal meal) {
+    public Meal save(Meal meal, int userId) {
+        log.info("save {}", meal);
         if (meal.isNew()) {
             meal.setId(counter.incrementAndGet());
             repository.put(meal.getId(), meal);
             return meal;
         }
-        // treat case: update, but absent in storage
-        return repository.computeIfPresent(meal.getId(), (id, oldMeal) -> meal);
+        if (meal.getUserId() == userId) {
+            repository.put(meal.getId(), meal);
+            return meal;
+        }
+        return null;
     }
 
     @Override
-    public void delete(int id) {
-        repository.remove(id);
+    public boolean delete(int id, int userId) {
+        log.info("delete {}", id);
+        Meal meal = repository.get(id);
+        if (meal == null) {
+            return false;
+        }
+        if (meal.getUserId() == userId) {
+            repository.remove(id);
+            return true;
+        }
+        return false;
     }
 
     @Override
-    public Meal get(int id) {
-        return repository.get(id);
+    public Meal get(int id, int userId) {
+        log.info("get {}", id);
+        Meal meal = repository.get(id);
+        if (meal == null) {
+            return null;
+        }
+        if (meal.getUserId() == userId) {
+            return meal;
+        }
+        return null;
     }
 
     @Override
-    public Collection<Meal> getAll() {
-        return repository.values();
+    public List<Meal> getAll() {
+        log.info("getAll");
+        return (List<Meal>) repository.values();
+    }
+
+    @Override
+    public List<Meal> getAllByUserId(final int userId) {
+        log.info("getAllByUserId {}", userId);
+        return repository.values()
+                .stream()
+                .filter(meal -> meal.getUserId() == userId)
+                .sorted((m1, m2) -> m2.getDateTime().compareTo(m1.getDateTime()))
+                .collect(Collectors.toList());
+    }
+
+    @Override
+    public List<MealWithExceed> getFiltered(final int userId, int caloriesPerDay, final LocalTime startTime, final LocalTime endTime) {
+        log.info("getFiltered {}", userId);
+        List<Meal> meals = getAllByUserId(userId);
+        return MealsUtil.getFilteredWithExceeded(meals, caloriesPerDay, startTime, endTime);
     }
 }
